+++ oxasl/reg.py
@@ -99,7 +99,7 @@ def get_motion_params(mat):
             mat[2, 1] - mat[1, 2],
             mat[0, 2] - mat[2, 0],
             mat[1, 0] - mat[0, 1],
-        ], dtype=np.float)
+        ], dtype=np.float32)
 
         # Rotation angle - note that we need to check the sign
         trace = np.trace(mat[:3, :3])
@@ -392,7 +392,7 @@ def transform(wsp, img, trans, ref, use_flirt=False, interp="trilinear", padding
         ret = fsl.applywarp(img, ref, out=fsl.LOAD, interp=interp, paddingsize=paddingsize, super=True, superlevel="a", log=wsp.fsllog, **kwargs)["out"]
     if mask:
         # Binarise mask images
-        ret = Image((ret.data > mask_thresh).astype(np.int), header=ret.header)
+        ret = Image((ret.data > mask_thresh).astype(np.int32), header=ret.header)
     return ret
 
 def reg_flirt(wsp, img, ref, initial_transform=None):
